requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-substrate-interface==1.1.0
asyncstdlib==3.13.1
attrs==25.3.0
base58==2.1.1
bittensor==9.4.0
bittensor-commit-reveal==0.4.0
bittensor-wallet==3.0.8
bt-decode==0.6.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
cytoolz==1.0.1
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
eth-hash==0.7.1
eth-typing==5.2.1
eth-utils==2.2.2
fastapi==0.110.3
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
langsmith==0.3.33
more-itertools==10.7.0
msgpack==1.1.0
msgpack-numpy-opentensor==0.5.0
multidict==6.4.3
munch==2.5.0
nest-asyncio==1.6.0
netaddr==1.3.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.76.0
orjson==3.10.16
packaging==25.0
password-strength==0.0.3.post2
platformdirs==4.3.7
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
py==1.11.0
py-bip39-bindings==0.1.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dotenv==1.1.0
python-statemachine==2.5.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rsa==4.9.1
scalecodec==1.2.11
sentry-sdk==2.26.1
setproctitle==1.3.5
setuptools==70.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.37.2
tenacity==9.0.0
toml==0.10.0
toolz==1.0.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
wandb==0.19.10
websockets==15.0.1
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
